+++ tests/testthat/test-calc_topological_roles.R
@@ -1,0 +1,9 @@
+
+
+test_that("calc_topological_roles", {
+  g <- netData[[1]]
+  set.seed(1432)
+  expect_snapshot(
+    calc_topological_roles(g,nsim=10,ncores=4)
+  )
+})