+++ NEWS.md
@@ -5,3 +5,9 @@ equivalent to a ecological transfer efficieny of 10%.
 Previusly they were `negative=-10, positive=.1, selfDamping=-1`.
 
 * Updated documentation with better explication of calc_interaction_intensity function
+
+# Version  0.6.31.0000
+
+* Updated documentation of calc_interaction_intensity function
+
+* Add @importFrom kSamples to calc_QSS_extinction_dif